src/fastapi/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pymongo==4.6.0
beanie==1.23.0
python-dotenv==1.0.0
pytest==7.3.5
httpx==0.25.0
gradio>=5.26.0
requests>=2.32.0
pymc-marketing>=0.13.1
pandas>=2.0.0
numpy>=1.22.0
matplotlib>=3.5.0
scikit-learn>=1.1.0
seaborn>=0.12.0
email-validator==2.1.0
motor==3.3.1

# Testing dependencies
pytest-asyncio==0.21.1
pytest-cov==4.1.0